requirements.txt
aiohttp==0.21.1
docker-py==1.8.0
ujson==1.35
